+++ src/main/java/umc/project/umark/domain/member/controller/MemberController.java
@@ -1,5 +1,8 @@
-package umc.project.umark.domain.member;
+package umc.project.umark.domain.member.controller;
 
+import umc.project.umark.domain.member.converter.MemberConverter;
+import umc.project.umark.domain.member.dto.MemberDto;
+import umc.project.umark.domain.member.service.MemberService;
 import umc.project.umark.global.exception.ApiResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.PostMapping;
