+++ src/main/java/umc/project/umark/domain/member/converter/MemberConverter.java
@@ -1,4 +1,7 @@
-package umc.project.umark.domain.member;
+package umc.project.umark.domain.member.converter;
+
+import umc.project.umark.domain.member.dto.MemberDto;
+import umc.project.umark.domain.member.entity.MemberStatus;
 
 public class MemberConverter {
