+++ src/main/java/umc/project/umark/domain/member/entity/MemberStatus.java
@@ -1,4 +1,4 @@
-package umc.project.umark.domain.member;
+package umc.project.umark.domain.member.entity;
 
 public enum MemberStatus {
     ACTIVE, INACTIVE, WITHDRAWN;
