+++ src/main/java/umc/project/umark/domain/term/entity/Term.java
@@ -2,7 +2,7 @@ package umc.project.umark.domain.term.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
-import umc.project.umark.domain.member.Member;
+import umc.project.umark.domain.member.entity.Member;
 import umc.project.umark.global.common.BaseEntity;
 
 @Getter
